# [sk] Statistics performance updates