# Summary
Updates to speed up statistics calculation. 

# Performance Metrics
Performance was evaluated on dataset with 100000 rows and 28 columns (total of 2.8 million entries) with numbers, categories, boolean, and text columns. All times are averaged over 100 trials. 
**Baseline**: 1.463 seconds
- No multithreading - 1.408 seconds (likely because statistics calculation is CPU bound rather than IO bound)
- Deduplicated null value calculation between null_value_rate and null_value_count - 1.362 seconds      
- Switched to inbuilt mean calculation - 1.344 seconds
- Switched from quantile to inbuilt median calculation - 1.323 seconds
- Refactored timeseries check to avoid manually calculating the null value rate - 1.285 seconds
- Optimizing longest null sequence:
    - Switched from pandas series to Numpy ndarray - 1.211 seconds
    - Switched to generator of null values (implicit running mean) - 1.04 seconds
- Removed duplicate max_null_sequence calculation - 0.91 seconds

**Final Time:** 0.91 seconds (~38% decrease)

# Tests
Correctness was confirmed by running the code on unit tests, and also through local testing.

cc:
@wangxiaoyou1993