@wangxiaoyou1993 2022-06-06
You can also reduce the duplicate `value_counts()` calculation.