@skunichetty 2022-06-07
I deduplicated the value counts call when computing the mode, reducing the runtime to 0.7568003163 when averaged over 100 trials.